Use HTTPS backends instead of SSL passthrough for Elasticsearch and Kibana
Elasticsearch supports custom certificates but Kibana can't
communicate with the internal cluser address of the Elasticsearch
service if it's configured to use a certificate issued by Let's
Encrypt.

It's possible to change the URL which Kibana uses to communicate
with Elasticsearch. But that _seem_ to require you also having
to configure a custom CA certificate and auth secrets.

Worth looking into in the future but this currently looks to be the
proposed solution:
https://github.com/elastic/cloud-on-k8s/issues/1010#issuecomment-499795679