Changed harbor installation to use certificates by cert-manager and provided secrets. This removes the problem that harbor always wanted to reinstall when running the helmfile again.
To achieve this a new commit to the harbor helm chart was needed.
This was only present on master, not any release.
So the chart was added to this repo. But it should be removed once a new version is released.

Currently the added secrets are hardcoded and should be randomized instead.
A trello card was added to fix this and other hardcoded/default secrets and passwords.