Add service cluster logging
This adds a logging solution to the service cluster. Fluentd is
used similarly to the workload cluster, however, instead of
sending logs to Elasticsearch the logs are forwarded to a Fluentd
aggregator.

The Fluentd aggregator is a single Fluentd deployment which outputs
all forwarded logs to an S3 bucket. Currently all logs are buffered
for 10 minutes on the aggregator until they are flushed to the S3
bucket.

To prevent any loss of log data the buffer is also persisted to a
persistent volume which is currently configured with 10GiB size.

Both the flush time and the volume storage could be parameterized
in the future. Though, a proper logging solution should probably
be implemented where the logs are more easily accessible and
searchable.