Disabled http logging for InfluxDB.
With http logging turned on we get pretty spammed by, in my opionin,
quite meaningless logs saying that the return code was this and that
etc. This would have been fine if the number of requests wasn't too
high. I think we get something like 30-100 requests/second.

I propse that we instead use Influx's prometheus endpoint to get http
metrics such as requests/second and number of status return codes.