apps sc: Added SLM policy.
This commit introduces Snapshot Lifecycle Management and it is the same
concept as ILM, but instead of managing indecies it manages snapshots.

Prior to this commit we took a snapshot by having a cronjob invoke the
snapshot api in elasticsearch. We did not have any retention set for the
snapshots that where taken.

Now SLM invokes the snapshot api once everyday at 1.30am and the default
retention is set to 30 days.