k8s: Update rke cluster naming.
This commit fixes the issue where all of our ck8s clusters had the same
name making it impossible to know which cluster you are accessing with
checking your environment variables.

Now the clusters will be named as follows:
${ENVIRONMENT_NAME}_${CLOUD_PROVIDER}{sc,wc}

This change should be possible to run on an already existing cluster, it
will create new(?) kubeconfigs that reflects the change in cluster
naming.