all: Add initial facade
This introduces a facade for CK8S which should serve as the
interface that CK8S users will be fronted with.

It mainly moves the logic from the config repositories and makes
all top-level configuration options required (except for those
which a user is not supposed to set when they are not used such
as OPSGENIE_HEARTBEAT_NAME).

It also replaces the use of Vault as secret management with SOPS.
Secrets can then safely be shared among collaborators by, for
example, commiting them to a shared git repository, since the
secrets are always encrypted at rest.

From now on the only environment variables required to be set in
the shell are:
CK8S_CONFIG_PATH
CK8S_CLOUD_PROVIDER *
CK8S_ENVIRONMENT_NAME *
SOPS_PGP_FP
TF_TOKEN *

* Only required during initialization.

Everything else is configured in configuration files located at
${CK8S_CONFIG_PATH}.