k8s: changed cni from canal to calico
Canal was initially used as cni since it's used by RKE.
However, the cni was troublesome to get working properly with the
kubeadm installation.
Calico was much easier to get started with.
Since there were no other reasons to why we are using canal, the
cni was changed to calico.