apps sc: Allow monitoring indices to be created.
This commit enables automatic index creation of kibana monitoring
indices. This allows for the activation of self-monitoring in kibana.
Previously, when one tried to activate self-monitoring nothing happened
because the index could not be created. It is obvious that kibana writes
to the index instead of explicitly creating the index before writing.