infra exo: Remove NFS server whitelisting
Since the NFS server was whitelisting specific worker IPs it was
required to reboot the NFS machine each time you added or removed
a worker machine for it to take effect.

For now we rely on the network being whitelisted by the security
groups, however whitelisting should be added back when we got
private networks in place.