tf aws: Fix extra_tags
ck8s_linux_amd64 (also called ckctl in the pipeline) initializes the
config file with *all* values, even default ones that the operator does
not or should not care about. Hence, when adding a new variable, it is
insufficient to add the default in the Terraform script, but this needs
to be also default-initialized in ckctl.

This script fixes a previously broken `extra_tags` contribution, that
omitted the default-initialization in the Go part.